// Frontend/src/app/manager/models/edit-personal.model.ts
export class PersonalEdit {
    personalId: number;
    currentFirstName: string;
    currentLastName: string;
    currentMiddleName : string;
    birthDate: Date;
    socialSecurityNumber: string;
    driversLicense: string;
    currentAddress1: string;
    currentAddress2: string;
    currentCity: string;
    currentCountry: string;
    currentZip: number;
    currentGender: string;
    currentPhoneNumber: string;
    currentPersonalEmail: string
    currentMaritalStatus: string;
    ethnicity: string;
    shareholderStatus: number;
    benefitPlanId: number;
}